// src/app/core/models/admEmpleados.model.ts
export interface admEmpleados {
    uiNumeroEmpleado: number;
    sNombreUsuario: string;
    sPuesto: string;
    sCorreoUsuario: string;
    uiUbicacion: string;
    sUEN: string;
    sEstatus: string;
    bInterno: boolean;
    bStatus:boolean;
    iUbicacion:number,
    // sNombreUbicacion?:string,
    uiIdUicacion?:number,
}

export class admEmpleados {

}

// public int uiNumeroEmpleado { get; set; }
// public string sNombreUsuario { get; set; }
// public string SPuesto { get; set; }
// public string sCorreoUsuario { get; set; }
// public string uiUbicacion { get; set; }
// public string sUEN { get; set; }
// public string sEstatus { get; set; }
// public bool bInterno { get; set; }